Update README.md instructions

+++ .dotfiles/init/scripts/gpu_drivers.sh
@@ -61,33 +61,32 @@ amd_drivers () {
     ##############################
 	# Enables overclocking below
 	##############################
-	yellow='\033[1;33m'
-	reset='\033[0m' # reset to default color
 	original_conf_file="$SCRIPT_DIR/system/etc/amd-overclock.original"
 	conf_file="/etc/amd-overclock.conf"
 
+	if [ ! -f "$conf_file" ]; then
+		cp -f $original_conf_file $conf_file
+		
+		# Comment out each variable in the config file
+		sed -i 's/^\(VOLTAGE_OFFSET=.*\)/#\1/' "$conf_file"
+		sed -i 's/^\(CORE_CLOCK=.*\)/#\1/' "$conf_file"
+		sed -i 's/^\(MEMORY_CLOCK=.*\)/#\1/' "$conf_file"
+		sed -i 's/^\(MAX_WATTS_POWERLIMIT=.*\)/#\1/' "$conf_file"
 
-	cp -f $original_conf_file $conf_file
-	
-	# Comment out each variable in the config file
-	sed -i 's/^\(VOLTAGE_OFFSET=.*\)/#\1/' "$conf_file"
-	sed -i 's/^\(CORE_CLOCK=.*\)/#\1/' "$conf_file"
-	sed -i 's/^\(MEMORY_CLOCK=.*\)/#\1/' "$conf_file"
-	sed -i 's/^\(MAX_WATTS_POWERLIMIT=.*\)/#\1/' "$conf_file"
+		clear_screen
+		printf "Overclock values will be commented out by default.\n\n"
+		printf "You will need to modify ${YELLOW}$conf_file${RESET}"
 
-	clear_screen
-	printf "Overclock values will be commented out by default.\n\n"
-	printf "You will need to modify ${yellow}$conf_file${reset}"
+		printf "\n\n"
+		read -p "Press enter to continue"
 
-	printf "\n\n"
-	read -p "Press enter to continue"
+		# Edit config file 
+	    pacman -Syu micro --ask 4 --needed
+		micro $conf_file
 
-	# Edit config file 
-    pacman -Syu micro --ask 4 --needed
-	micro $conf_file
-
-	# Enable service
-	systemctl enable amd-overclock.service
+		# Enable service
+		systemctl enable amd-overclock.service
+	fi
 }
 
 intel_drivers () {
